update fuchsia device discovery to use tool mdns
